+++ engine/src/main/java/com/etherblood/luna/engine/MilliHealth.java
@@ -1,6 +1,6 @@
 package com.etherblood.luna.engine;
 
 public record MilliHealth(
-        int value
+        long value
 ) {
 }
