+++ engine/src/main/java/com/etherblood/luna/engine/Speed.java
@@ -1,6 +1,6 @@
 package com.etherblood.luna.engine;
 
 public record Speed(
-        int milliMetresPerFrame
+        long milliMetresPerFrame
 ) {
 }
